requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
celery
redis
openai
pymupdf
python-dotenv